+++ src/injected/drawing-controller.ts
@@ -15,7 +15,7 @@ import {
     HTMLIFrameResult,
 } from './frameCommunicators/html-element-axe-results-helper';
 import { FrameMessageResponseCallback } from './frameCommunicators/window-message-handler';
-import { RegisterDrawer } from './visualization-type-drawer-registrator';
+import { RegisterDrawer } from './visualization-type-drawer-registrar';
 import { Drawer } from './visualization/drawer';
 
 export interface VisualizationWindowMessage {
