+++ src/tests/unit/tests/injected/visualization-type-drawer-registrar.test.ts
@@ -6,16 +6,16 @@ import { Requirement } from '../../../../assessments/types/requirement';
 import { VisualizationConfiguration } from '../../../../common/configs/visualization-configuration';
 import { VisualizationConfigurationFactory } from '../../../../common/configs/visualization-configuration-factory';
 import { VisualizationType } from '../../../../common/types/visualization-type';
-import { RegisterDrawer, VisualizationTypeDrawerRegistrator } from '../../../../injected/visualization-type-drawer-registrator';
+import { RegisterDrawer, VisualizationTypeDrawerRegistrar } from '../../../../injected/visualization-type-drawer-registrar';
 import { Drawer } from '../../../../injected/visualization/drawer';
 import { DrawerProvider } from '../../../../injected/visualization/drawer-provider';
 
-describe('VisualizationTypeDrawerRegistrator', () => {
+describe('VisualizationTypeDrawerRegistrar', () => {
     let registerDrawerMock: IMock<RegisterDrawer>;
     let visualizationConfigFactoryMock: IMock<VisualizationConfigurationFactory>;
     let assessmentProviderMock: IMock<AssessmentsProvider>;
     let drawerProviderMock: IMock<DrawerProvider>;
-    let testSubject: VisualizationTypeDrawerRegistrator;
+    let testSubject: VisualizationTypeDrawerRegistrar;
     let typeStub: VisualizationType;
     let configMock: IMock<VisualizationConfiguration>;
     let identifierStub: string;
@@ -31,7 +31,7 @@ describe('VisualizationTypeDrawerRegistrator', () => {
         identifierStub = 'some id';
         drawerStub = {} as Drawer;
 
-        testSubject = new VisualizationTypeDrawerRegistrator(
+        testSubject = new VisualizationTypeDrawerRegistrar(
             registerDrawerMock.object,
             visualizationConfigFactoryMock.object,
             assessmentProviderMock.object,
